Replaced doubles by Reals, fixed two bugs
* Replaced all doubles by Reals where necessary
* Replaced constructor
    TFactor<T>::TFactor( const VarSet& vars, TIterator begin )
  by two constructors
    TFactor<T>::TFactor( const VarSet& vars, const T* p )
    TFactor<T>::TFactor( const VarSet& vars, const std::vector<S> &x )
* Fixed bugs (min and max were reversed) in
    TFactor<T> min( const TFactor<T> &, const TFactor<T> & )
    TFactor<T> max( const TFactor<T> &, const TFactor<T> & )

+++ utils/fginfo.cpp
@@ -108,13 +108,13 @@ int main( int argc, char *argv[] ) {
             cout << "Treewidth:           " << tw.first << endl;
             cout << "Largest cluster for JTree has " << tw.second << " states " << endl;
         }
-        double stsp = 1.0;
+        long double stsp = 1.0;
         for( size_t i = 0; i < fg.nrVars(); i++ )
             stsp *= fg.var(i).states();
         cout << "Total state space:   " << stsp << endl;
 
-        double cavsum_lcbp = 0.0;
-        double cavsum_lcbp2 = 0.0;
+        long double cavsum_lcbp = 0.0;
+        long double cavsum_lcbp2 = 0.0;
         size_t max_Delta_size = 0;
         map<size_t,size_t> cavsizes;
         for( size_t i = 0; i < fg.nrVars(); i++ ) {
